refactor(cli): extract buildProgram factory for doc generation
